backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cbor2==5.7.1
cffi==2.0.0
channels==4.3.2
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
zope.interface==8.1.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.35.0
channels-redis==4.1.0
